backend/requirements.txt
boto3==1.33.13
botocore==1.33.13
click==8.1.8
colorama==0.4.6
cycler==0.11.0
et-xmlfile==1.1.0
Flask==2.2.5
Flask-Cors==5.0.0
fonttools==4.38.0
importlib-metadata==6.7.0
itsdangerous==2.1.2
jinja2==3.1.6
jmespath==1.0.1
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.5.3
numpy==1.21.6
openpyxl==3.1.3
packaging==24.0
pandas==1.1.5
Pillow==9.5.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
pytz==2025.1
s3transfer==0.8.2
seaborn==0.12.2
six==1.17.0
typing-extensions==4.7.1
urllib3==1.26.20
Werkzeug==2.2.3
xlrd==2.0.1
zipp==3.15.0
